Update README.md
Data collection in Python refers to the process of gathering data from various sources such as databases, APIs, sensors, web pages, and more, and then storing or processing it. When collecting data through web scraping or making repeated API requests, your IP can be blocked by the server if it detects unusual or excessive traffic. To prevent this, several strategies can be employed in Python.